Renamed to js_email_password_auth.hpp
