+++ binding.gyp
@@ -27,7 +27,7 @@
             "src/js_app_credentials.hpp",
             "src/js_user.hpp",
             "src/js_network_transport.hpp",
-            "src/js_email_password_provider.hpp",
+            "src/js_email_password_auth.hpp",
             "src/node/sync_logger.cpp",
             "src/node/sync_logger.hpp",
           ]
